Refactor timeout management by introducing ThreadPoolTimeoutExecutor
### Summary
- Removed the old timeout executor import from `AbstractOperation` and replaced it with the new `ThreadPoolTimeoutExecutor`.
- Added the `ThreadPoolTimeoutExecutor` class to manage operation timeouts using a shared thread pool.
- Updated references in `SessionManager` and test suite to align with the new executor structure.

### Changes
- Introduced `ThreadPoolTimeoutExecutor` for better resource management of operation timeouts.
- Refactored imports in relevant files to ensure proper usage of the new executor.
- Enhanced unit tests to validate the functionality of the new timeout management approach.

This update improves the clarity and efficiency of timeout handling in Kyuubi.
