+++ kyuubi-common/src/main/scala/org/apache/kyuubi/session/SessionManager.scala
@@ -29,7 +29,7 @@ import org.apache.kyuubi.{KyuubiSQLException, Utils}
 import org.apache.kyuubi.config.KyuubiConf
 import org.apache.kyuubi.config.KyuubiConf._
 import org.apache.kyuubi.operation.OperationManager
-import org.apache.kyuubi.operation.timeout.ThreadPoolTimeoutExecutor
+import org.apache.kyuubi.operation.ThreadPoolTimeoutExecutor
 import org.apache.kyuubi.service.CompositeService
 import org.apache.kyuubi.shaded.hive.service.rpc.thrift.TProtocolVersion
 import org.apache.kyuubi.util.ThreadUtils
